+++ src/props/vocabulary/index.ts
@@ -28,6 +28,7 @@ export type {
 export type {
   TitleProp,
   SubtitleProp,
+  StatusProp,
   DescriptionProp,
   ExtraProp,
   FooterProp,